pyproject.toml
[tool.poetry]
name = "extern-2"
version = "0.1.0"
description = "Central University 2nd project for extern"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
annotated-types = "0.7.0"
anyio = "4.6.2.post1"
certifi = "2024.8.30"
charset-normalizer = "3.4.0"
click = "8.1.7"
fastapi = "0.115.3"
h11 = "0.14.0"
idna = "3.10"
jinja2 = "3.1.4"
markupsafe = "3.0.2"
pydantic = "2.9.2"
pydantic-core = "2.23.4"
python-dotenv = "1.0.1"
python-multipart = "0.0.12"
requests = "2.32.3"
sniffio = "1.3.1"
starlette = "0.41.0"
typing-extensions = "4.12.2"
urllib3 = "2.2.3"
uvicorn = "0.32.0"
dash = "^2.18.1"
dash-core-components = "^2.0.0"
dash-html-components = "^2.0.0"
plotly = "^5.24.1"
gunicorn = "^23.0.0"
aiogram = "^3.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
